+++ trinity/buffer/pipelines/__init__.py
@@ -1,0 +1,5 @@
+from trinity.buffer.pipelines.experience_pipeline import ExperiencePipeline
+
+__all__ = [
+    "ExperiencePipeline",
+]
