--- trinity/data/pipelines/__init__.py
@@ -1,5 +1,0 @@
-from trinity.data.pipelines.experience_pipeline import ExperiencePipeline
-
-__all__ = [
-    "ExperiencePipeline",
-]
